fix: rename suspense function
