+++ src/common/index.ts
@@ -1,2 +1,2 @@
 export { PromiseCache as CacheService } from './PromiseCache';
-export { suspense } from './suspense';
+export { lazyData } from './lazyData';
