+++ src/common/lazyData.spec.ts
@@ -1,10 +1,10 @@
-import { StatefulPromise, suspense } from './suspense';
+import { StatefulPromise, lazyData } from './lazyData';
 
-describe('suspense', () => {
+describe('lazyData', () => {
   it('should throw the promise and set its status to pending if it is not resolved/rejected', () => {
     const p = new Promise(() => {});
     p.then = jest.fn();
-    const cb = () => suspense(p);
+    const cb = () => lazyData(p);
 
     expect(cb).toThrow();
     expect(p.then).toBeCalled();
@@ -18,12 +18,12 @@ describe('suspense', () => {
   it('should throw rejected value if the promise is rejected', () => {
     const er = new Error('Rejected');
     const p = { status: 'rejected', error: er } as StatefulPromise<string>;
-    const cb = () => suspense(p);
+    const cb = () => lazyData(p);
     expect(cb).toThrow(er);
   });
 
   it('should return resolved value if promise is resolved', () => {
     const p = { status: 'fulfilled', value: 'data' } as StatefulPromise<string>;
-    expect(suspense(p)).toBe('data');
+    expect(lazyData(p)).toBe('data');
   });
 });
