+++ src/common/lazyData.ts
@@ -16,7 +16,7 @@ export interface StatefulPromise<T> extends Promise<T> {
  * @throws if the promise is not resolved, throw the promise.
  * @throws if the promise is rejected, throw the rejected value.
  */
-export function suspense<T>(promise: StatefulPromise<T>) {
+export function lazyData<T>(promise: StatefulPromise<T>) {
   // update promise's attributes when resolved/rejected
   if (!promise.status) {
     /* eslint-disable no-param-reassign */
